Remove codalabworker package
