+++ codalab/rest/bundle_actions.py
@@ -5,7 +5,7 @@ from codalab.lib.bundle_action import BundleAction
 from codalab.objects.permission import check_bundles_have_all_permission
 from codalab.rest.schemas import BundleActionSchema
 from codalab.server.authenticated_plugin import AuthenticatedPlugin
-from codalabworker.bundle_state import State
+from codalab.worker.bundle_state import State
 
 
 @post('/bundle-actions', apply=AuthenticatedPlugin())
