+++ codalab/rest/workers.py
@@ -184,21 +184,3 @@ def workers_info():
         worker["gpus_in_use"] = sum(bundle.metadata.request_gpus for bundle in running_bundles)
 
     return {"data": data}
-
-
-@get("/workers/code.tar.gz", name="worker_download_code")
-def code():
-    """
-    Returns .tar.gz archive containing the code of the worker.
-    """
-    response.set_header("Content-Disposition", 'attachment; filename="code.tar.gz"')
-    response.set_header("Content-Type", "application/gzip")
-    codalab_cli = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-    code_dir = os.path.join(codalab_cli, "worker", "codalabworker")
-    args = ["tar", "czf", "-", "-C", code_dir]
-    for filename in os.listdir(code_dir):
-        if filename.endswith(".py") or filename.endswith(".sh"):
-            args.append(filename)
-    proc = subprocess.Popen(args, stdout=subprocess.PIPE)
-    result = proc.stdout.read()
-    return result
