+++ setup.py
@@ -7,19 +7,8 @@ CODALAB_VERSION = "0.4.0"
 if int(setuptools.__version__.split('.')[0]) < 25:
     print(
         "WARNING: Please upgrade setuptools to a newer version, otherwise installation may break. "
-        "Recommended command: `pip install -U setuptools`"
+        "Recommended command: `pip3 install -U setuptools`"
     )
-
-
-def get_requirements(*requirements_file_paths):
-    requirements = ['codalabworker>={}'.format(CODALAB_VERSION)]
-    for requirements_file_path in requirements_file_paths:
-        with open(requirements_file_path) as requirements_file:
-            for line in requirements_file:
-                if line[0:2] != '-r' and line.find('git') == -1:
-                    requirements.append(line.strip())
-    return requirements
-
 
 setup(
     name='codalab',
@@ -42,9 +31,12 @@ setup(
     ],
     python_requires='~=3.6',
     include_package_data=True,
-    install_requires=get_requirements('worker/requirements.txt'),
     entry_points={
-        'console_scripts': ['cl=codalab.bin.cl:main', 'codalab-service=codalab_service:main']
+        'console_scripts': [
+            'cl=codalab.bin.cl:main',
+            'codalab-service=codalab_service:main',
+            'cl-worker=codalab.worker.main:main',
+        ]
     },
     zip_safe=False,
 ),
