+++ tests/lib/canonicalize_test.py
@@ -4,7 +4,7 @@ import unittest
 
 from codalab.common import UsageError
 from codalab.lib import canonicalize, spec_util
-from codalabworker.bundle_state import State
+from codalab.worker.bundle_state import State
 
 
 class CanonicalizeTest(unittest.TestCase):
