--- worker/requirements.txt
@@ -1,12 +1,0 @@
-bottle==0.12.9
-docker==3.7.0
-marshmallow-jsonapi==0.15.1
-marshmallow==2.6.1
-setuptools>=40.0.0
-argcomplete==1.9.4
-PyYAML==5.1
-psutil==5.4.6
-six==1.11.0
-SQLAlchemy==1.0.8
-watchdog==0.8.3
-fusepy==2.0.4
